fix(ios): 修复 4 个编译错误 + 4 个运行时风险
Critical 编译错误修复:
- C1: project.yml 排除 Widget/ 目录，避免双 @main 冲突
  Widget 文件归入独立 Widget Extension Target
- C2: KeychainManager.swift 添加 import CommonCrypto
  CCKeyDerivationPBKDF 等符号现在可正确解析
- C3: PluginExecutor.execute() 的 executeBroadcast 调用添加 await
- C4: ContentView selectedTab 改为 SidebarTab? (可选类型)
  List(_:selection:) 要求 Binding<SelectionValue?>

Warning 运行时风险修复:
- W3: SessionsView 删除/创建会话同步 SwiftData 持久化
  (deleteMessages + deleteSession + saveSession)
- W5: SettingsView temperature 类型 Double → Float
  (与 AgentConfig.temperature 一致)
- W6: McpView .autocapitalization(.none) → .textInputAutocapitalization(.disabled)
- W11: AgentConfigEntity 新增 protocolType 字段持久化
  AgentConfig 同步添加 protocolType 属性
  toDomain/fromDomain 双向映射完整

+++ ios/AgentControlCenter/Features/SettingsView.swift
@@ -11,7 +11,7 @@ struct SettingsView: View {
 
     // ============ Agent 默认配置 ============
     @AppStorage("defaultModel") private var defaultModel: String = "gpt-4"
-    @AppStorage("temperature") private var temperature: Double = 0.7
+    @AppStorage("temperature") private var temperature: Float = 0.7
     @AppStorage("maxTokens") private var maxTokens: Int = 4096
 
     // ============ 字体大小 ============
